% Generated by r6-generator-maven-plugin: do not edit by hand

\name{delete_slides}
\alias{delete_slides}
\title{Deletes a google slides by name.}
\usage{delete_slides(
	docName,
	areYouSure,
	tokenDirectory,
	disabled
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{docName}{docName - the name of a document to delete. must be an exact and unique match. - (java expects a RCharacter)}
	\item{areYouSure}{areYouSure - a boolean check. - (defaulting to `utils::askYesNo(paste0('Are you sure ...`) - (java expects a RLogical)}
	\item{tokenDirectory}{tokenDirectory - (defaulting to `.tokenDirectory()`) - (java expects a RCharacter)}
	\item{disabled}{disabled - (defaulting to `getOption('roogledocs.disabled',FALSE)`) - (java expects a RLogical)}
	\if{html}{\out{</div>}}
}
\value{
void: nothing, called for side efffects

}
\description{
no description
}
\keyword{java api}


