% Generated by r6-generator-maven-plugin: do not edit by hand

\name{doc_by_id}
\alias{doc_by_id}
\title{Select a document by its share url or id.}
\usage{doc_by_id(
	shareUrlOrDocId,
	tokenDirectory=.tokenDirectory(),
	disabled=getOption('roogledocs.disabled',FALSE)
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{shareUrlOrDocId}{shareUrlOrDocId the url from clicking a share button in google docs or an id from searchForDocuments() method - (java expects a String)}
	\item{tokenDirectory}{tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to .tokenDirectory()) - (java expects a String)}
	\item{disabled}{disabled a flag to switch roogledocs off (on a document by document basis, for testing or development. This can be set globally with `options('roogledocs.disabled'=TRUE)` - (defaulting to getOption('roogledocs.disabled',FALSE)) - (java expects a boolean)}
	\if{html}{\out{</div>}}
}
\value{
R6 RoogleDocs object: itself - a fluent method
}
\description{
no description
}
\keyword{java api}


