% Generated by r6-generator-maven-plugin: do not edit by hand

\name{doc_from_template}
\alias{doc_from_template}
\title{Search for a document by name or create one if missing.}
\usage{doc_from_template(
	title,
	templateUri,
	tokenDirectory=.tokenDirectory(),
	disabled=getOption('roogledocs.disabled',FALSE)
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{title}{title a document title. If there is an exact match in google drive then that document will be used
otherwise a new one will be created. - (java expects a String)}
	\item{templateUri}{templateUri the share link (or document id) of a template google document - (java expects a String)}
	\item{tokenDirectory}{tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to .tokenDirectory()) - (java expects a String)}
	\item{disabled}{disabled a flag to switch roogledocs off (on a document by document basis, for testing or development. This can be set globally with `options('roogledocs.disabled'=TRUE)` - (defaulting to getOption('roogledocs.disabled',FALSE)) - (java expects a boolean)}
	\if{html}{\out{</div>}}
}
\value{
R6 RoogleDocs object: itself - a fluent method
}
\description{
no description
}
\keyword{java api}


