% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_maven.R
\name{.as.coordinates}
\alias{.as.coordinates}
\title{Maven coordinates}
\usage{
.as.coordinates(groupId, artifactId, version, ...)
}
\arguments{
\item{groupId}{the maven `groupId`}

\item{artifactId}{the maven `artifactId`}

\item{version}{the maven version}

\item{...}{other parameters ignored apart from `packaging`
(one of `jar`,`war`,`pom` or `ejb`) and `classifier`
(one of `tests`, `client`, `sources`, `javadoc`, `jar-with-dependencies`,
or `src`)}
}
\value{
a coordinates object containing the Maven artifact coordinates
}
\description{
Maven coordinates
}
\keyword{internal}
