% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_rfuture.R
\name{.background_cancel}
\alias{.background_cancel}
\title{Async cancel one}
\usage{
.background_cancel(id)
}
\arguments{
\item{id}{the id from `.background_status()`}
}
\value{
nothing, called for side effects
}
\description{
Cancel an asynchronous operation by id.
}
