% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_rfuture.R
\name{.background_get_by_id}
\alias{.background_get_by_id}
\title{Get a RFuture by id}
\usage{
.background_get_by_id(id)
}
\arguments{
\item{id}{the id from `.background_status()`}
}
\value{
an RFuture
}
\description{
Get a background task from the `.background_status()` list and wrap it in 
an R6 RFuture class. This
}
