% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_rfuture.R
\name{.background_status}
\alias{.background_status}
\title{Async function status}
\usage{
.background_status()
}
\value{
a maybe empty dataframe with `id` and `status` columns
}
\description{
Print any asynchronous java function calls that are in progress, complete or
cancelled. The function returns a thread id which may be used to retrieve the
thread itself
}
