% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_maven.R
\name{.execute_maven}
\alias{.execute_maven}
\title{Executes a maven goal}
\usage{
.execute_maven(
  goal,
  opts = c(),
  pom_path = NULL,
  quiet = .quietly(verbose),
  debug = .debug(verbose),
  verbose = c("normal", "debug", "quiet"),
  require_jdk = FALSE,
  settings = .settings_path(),
  ...
)
}
\arguments{
\item{goal}{the goal of the `mvn` command ( can be multiple ) e.g.
`c("clean","compile")`}

\item{opts}{provided options in the form
`c("-Doption1=value2","-Doption2=value2")`}

\item{pom_path}{optional. the path to a `pom.xml` file for goals that need
one.}

\item{quiet}{should output from maven be suppressed? (`-q` flag)}

\item{debug}{should output from maven be verbose? (`-X` flag)}

\item{verbose}{how much output from maven, one of "normal", "quiet", "debug"}

\item{require_jdk}{does the goal you are executing require a `JDK` (e.g.
compilation does, fetching artifacts and calculating class path does not)}

\item{settings}{the path to a `settings.xml` file controlling Maven. The
default is a configuration with a local repository in the `rmaven` cache
directory (and not the Java maven repository).}

\item{...}{non-empty named parameters are passed to maven as options in the
form `-Dname=value`}
}
\value{
nothing, invisibly
}
\description{
Maven goals are defined either as life-cycle goals (e.g. "clean", "compile")
or as plugin goals (e.g. "help:system"). Some Maven goals may be executed
without a `pom.xml` file, others require one. Some maven goals (e.g.
compilation) require the use of a `JDK`.
}
\keyword{internal}
