% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_maven.R
\name{.fetch_artifact}
\alias{.fetch_artifact}
\title{Fetch an artifact, and dependencies, into the local `.m2` repository}
\usage{
.fetch_artifact(
  groupId = NULL,
  artifactId = NULL,
  version = NULL,
  ...,
  coordinates = NULL,
  artifact = NULL,
  repoUrl = .default_repos(),
  nocache = FALSE,
  verbose = c("normal", "quiet", "debug")
)
}
\arguments{
\item{groupId}{optional, the maven `groupId`,}

\item{artifactId}{optional, the maven `artifactId`,}

\item{version}{optional, the maven version,}

\item{...}{other maven coordinates such as classifier or packaging}

\item{coordinates}{optional, but if not supplied `groupId` and `artifactId`
must be, coordinates as a coordinates object (see `as.coordinates()`)}

\item{artifact}{optional, coordinates as an artifact string
`groupId:artifactId:version[:packaging[:classifier]]` string}

\item{repoUrl}{the URLs of the repositories to check (defaults to Maven
central, 'Sonatype' snapshots and 'jitpack', defined in
`options("rmaven.default_repos"))`}

\item{nocache}{normally artifacts are only fetched if required, `nocache`
forces fetching}

\item{verbose}{how much output from maven, one of "normal", "quiet", "debug"}
}
\value{
the path of the artifact within the local maven cache
}
\description{
This can be used to get a JAR file from the maven repositories into a
local `.m2` repository. The local path is made available for importing it into
the `rJava` classpath for example.
}
\keyword{internal}
