% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_maven.R
\name{.get_repository_location}
\alias{.get_repository_location}
\title{Get the location of the Maven repository}
\usage{
.get_repository_location(settings_path = .settings_path())
}
\arguments{
\item{settings_path}{the file path of the `settings.xml` to update (generally
the supplied default is what you want to use)}
}
\value{
the location of the maven repository
}
\description{
In general this function is mainly for internal use but maybe handy for
debugging. The maven repository location can be defined by
`set_repository_location(...)` or through the option
`options("rmaven.m2.repository"=...)` option but defaults to a
`.m2/repository` directory in the `rmaven` cache directory. This is not the
default location for Maven when used from Java writing to the default Maven
directory in user space is forbidden by CRAN policies. The result of this is
that `rmaven` will have to unnecessarily download additional copies of Java
libraries, onto the users computer and cannot re-use already cached copies.
Also Maven wont be able to pick up jar files locally installed through
standard Java tooling, unless the default CRAN approved location is overridden
using `set_repository_location("~/.m2/repository")`.
This is more of an issue for developers rather than users.
}
\keyword{internal}
