% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_maven.R
\name{.print.coordinates}
\alias{.print.coordinates}
\title{Prints a coordinates object}
\usage{
.print.coordinates(x, ...)
}
\arguments{
\item{x}{a maven coordinates object}

\item{...}{ignored}
}
\description{
Prints a coordinates object
}
\keyword{internal}
