% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_maven.R
\name{.start_jvm}
\alias{.start_jvm}
\title{Start an `rJava` `JVM` with or without debugging options}
\usage{
.start_jvm(
  debug = FALSE,
  quiet = getOption("rmaven.quiet", TRUE),
  max_heap = NULL,
  thread_stack = NULL,
  ...
)
}
\arguments{
\item{debug}{turn on debugging}

\item{quiet}{don't report messages (defaults to `getOption("rmaven.quiet")`
or TRUE)}

\item{max_heap}{optional. if a string like `"2048m"` the `-Xmx` option value
to start the `JVM` - if a string like `"75%"` the `-XX:MaxRAMPercentage`,
if a numeric - number of megabytes.}

\item{thread_stack}{optional. sensible values range from '1m' to '128m' (max
is '1g'). Can be important with deeply nested structures.}

\item{...}{any other named parameters are passed as `-name=value`}
}
\value{
nothing - called for side effects
}
\description{
This does not do anything if the `JVM` has already been started. Otherwise
starts the JVM via `rJava` with a set of options Additional JVM options
(beyond debugging) can be set with the
`options("java.parameters"=c("-Xprof","-Xrunhprof"))`
}
\keyword{internal}
