% Generated by r6-generator-maven-plugin: do not edit by hand

\name{reauth}
\alias{reauth}
\title{Re-authenticate roogledocs library}
\usage{reauth(
	tokenDirectory
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{tokenDirectory}{tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to `.tokenDirectory()`) - (java expects a String)}
	\if{html}{\out{</div>}}
}
\value{
R6 RoogleDocs object: a new RoogleDocs instance without an active document
}
\description{
Re-authenticate the service deleting the existing OAuth tokens may be helpful if there is some problem. 

Generally this is only be needed if  
application permission updates are needed in which case the directory can be manually deleted anyway,
or if you want to switch google user without using a different tokenDirectory.
}
\keyword{java api}


