% Generated by r6-generator-maven-plugin: do not edit by hand
\docType{package}
\name{roogledocs-package}
\alias{roogledocs}
\alias{roogledocs-package}
\title{roogledocs: R Wrapper For Googledocs Java Library}
\description{
% \if{html}{\figure{logo.png}{options: align='right'}}
Programmatically substitute images, data and tables into a google doc. R library to perform limited interactions with google docs (and maybe one day slides) in R via the Java API library. The purpose being to support google docs as a platform for interactive development and documentation of data analysis in R for scientific publication, although it is not limited to this purpose. The workflow supported is a parallel documentation and analysis where a team of people are working collaboratively on documentation, whilst at the same time analysis is being performed and results updated repeatedly as a result of new data. In this environment updating numeric results, tabular data and figures in word documents manually becomes annoying. With roogledocs you can automate this a bit like a RMarkdown document, but with the added benefit that the content can be updated independently of the analysis, by the wider team.

Version: 0.1.0

Classes:
\itemize{ 
	\item \code{JavaApi}
	\item \code{RoogleDocs}
}
}

\author{
\itemize{ 
   \item Rob Challen rob.challen@bristol.ac.uk 0000-0002-5504-7768
}
}

\keyword{java api}