% Generated by r6-generator-maven-plugin: do not edit by hand

\name{search_for_documents}
\alias{search_for_documents}
\title{Search for documents with the given title}
\usage{search_for_documents(
	titleMatch,
	tokenDirectory
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{titleMatch}{titleMatch a string to be searched for as an approximate match. All results will be retrieved with document ids. - (java expects a RCharacter)}
	\item{tokenDirectory}{tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to `.tokenDirectory()`) - (java expects a RCharacter)}
	\if{html}{\out{</div>}}
}
\value{
RDataframe: a dataframe containing id and name columns

}
\description{
no description
}
\keyword{java api}


