% Generated by r6-generator-maven-plugin: do not edit by hand

\name{slides_by_id}
\alias{slides_by_id}
\title{Get a document by id or sharing link.}
\usage{slides_by_id(
	shareUrlOrDocId,
	tokenDirectory,
	disabled
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{shareUrlOrDocId}{shareUrlOrDocId the url from clicking a share button in google slides or an id from searchForDocuments() method - (java expects a RCharacter)}
	\item{tokenDirectory}{tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to `.tokenDirectory()`) - (java expects a RCharacter)}
	\item{disabled}{disabled a flag to switch roogledocs off (on a document by document basis, for testing or development. This can be set globally with `options('roogledocs.disabled'=TRUE)` - (defaulting to `getOption('roogledocs.disabled',FALSE)`) - (java expects a RLogical)}
	\if{html}{\out{</div>}}
}
\value{
R6 RoogleSlides object: itself - a fluent method

}
\description{
no description
}
\keyword{java api}


