% Generated by r6-generator-maven-plugin: do not edit by hand

\name{slides_by_name}
\alias{slides_by_name}
\title{Get a document by name or create a blank document if missing.}
\usage{slides_by_name(
	title,
	tokenDirectory,
	disabled
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{title}{title a document title. If there is an exact match in google drive then that document will be used - (java expects a String)}
	\item{tokenDirectory}{tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to `.tokenDirectory()`) - (java expects a String)}
	\item{disabled}{disabled a flag to switch roogledocs off (on a document by document basis, for testing or development. This can be set globally with `options('roogledocs.disabled'=TRUE)` - (defaulting to `getOption('roogledocs.disabled',FALSE)`) - (java expects a boolean)}
	\if{html}{\out{</div>}}
}
\value{
R6 RoogleSlides object: itself - a fluent method
}
\description{
no description
}
\keyword{java api}


