% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\docType{data}
\name{std_size}
\alias{std_size}
\title{Standard image and paper sizes}
\format{
A list with width and height in inches
}
\usage{
std_size
}
\description{
The width and height of images to fit scientific publication standards.
}
\concept{output}
\keyword{datasets}
