% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-roogledocs.R
\name{with_project}
\alias{with_project}
\title{Set the main project name for a session.}
\usage{
with_project(name, version = "v0.01")
}
\arguments{
\item{name}{The main project name. Forms part of the googledocs or slides
filename.}

\item{version}{A version identifier. defaults to `"v0.01"`}
}
\value{
Nothing. Called for side effects. Does not check that document
  exists. If it does not exist it will be created on first use.
}
\description{
A project consists of a main document and optional supplementary files with
possible presentation versions. The associated google docs and slides will be
named with a versioned title based on the name parameter,
e.g. `"<name> <version>"`, `"<name> supplementary 1 <version>"`,
`"<name> slides <version>"`...
}
\examples{
if (FALSE) with_project("Test project")
}
\keyword{internal}
