% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-roogledocs.R
\name{write-to-docs}
\alias{write-to-docs}
\alias{project_doc}
\alias{write_value}
\alias{write_plot}
\alias{write_png}
\alias{write_table}
\alias{update_citations}
\title{Simpler interface to `roogledocs`}
\usage{
project_doc(supp = NA)

write_value(value, name = deparse(substitute(value)), supp = NA)

write_plot(p, name = deparse(substitute(p)), size = std_size$half, supp = NA)

write_png(png, name = deparse(substitute(png)), supp = NA)

write_table(
  t,
  name = deparse(substitute(t)),
  size = std_size$full,
  colwidths = rep(1, ncol(t)),
  supp = NA
)

update_citations(bibfile, style, supp = NA)
}
\arguments{
\item{supp}{use a supplementary? either a number e.g. 1, 2, 3... or "slides"}

\item{value}{A formatted text snippet.}

\item{name}{the tag name to replace it should be in the google doc as `\{\{tag\}\}`}

\item{p}{A ggplot object.}

\item{size}{a `std_size` object with width and height in inches.}

\item{png}{A path to a png image}

\item{t}{a huxtable table}

\item{colwidths}{an set of relative column widths, which will be made relative
to the total table width.}

\item{bibfile}{the location of the bibtex file}

\item{style}{a CSL (see `roogledocs::citation_styles()`)}
}
\value{
the written value invisibly, or the `roogledocs` document object
  for a project document.
}
\description{
A simple way to write to a document, supplementary or slides, without having to
manage individual document lifecycle, or learn a lot of . This is designed for the most usual
use case of writing a single manuscript with possibly associated slides and or
supplementary materials in the context of one R session. It allows a document
to be selected and written to iteratively.
}
\details{
This set of functions can be used even if `roogledocs` is not installed. In
this case the functions will short circuit and do nothing. `roogledocs` may
also be globally disabled.

These functions can be imported into a project with
`devtools::use_standalone("terminological/roogledocs")` to allow optional
simple `roogledocs` use.
}
\section{Functions}{
\itemize{
\item \code{project_doc()}: Get a document in the current project.

\item \code{write_value()}: Write a value to the current manuscript

\item \code{write_plot()}: Write a plot to a document in the current project

\item \code{write_png()}: Write a png to a document in the current project

\item \code{write_table()}: Write a table to a document in the current project

\item \code{update_citations()}: update the citations in the current manuscript

}}
\examples{
if (FALSE) {
  options(roogledocs.disabled=FALSE)
  with_project("Test project")
  tag1 = "some value"
  write_value(tag1)
  figure1 = ggplot2::ggplot()
  write_plot(figure1)
  table1 = huxtable::as.hux()
  write_table(table1)
  bibfile = "/path/to/my/bibfile"
  update_citations(bibfile,"vancouver")
}
}
\seealso{
[with_project()]
}
\keyword{internal}
