% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_rfuture.R
\name{.background_tidy_up}
\alias{.background_tidy_up}
\title{Tidy up completed async operations}
\usage{
.background_tidy_up()
}
\value{
nothing, called for side effects
}
\description{
Remove all processed and cancelled aync operations from the status list.
This will potentially free up some system resources
}
