% Generated by r6-generator-maven-plugin: do not edit by hand

\name{RoogleDocs}

\alias{RoogleDocs}

\title{RoogleDocs}

\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{tokenDirectory}{tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to `.tokenDirectory()`) - (java expects a String)}
	\item{disabled}{disabled a flag to switch roogledocs off (on a document by document basis, for testing or development. This can be set globally with `options('roogledocs.disabled'=TRUE)` - (defaulting to `getOption('roogledocs.disabled',FALSE)`) - (java expects a boolean)}
	\if{html}{\out{</div>}}
}


\description{
Programmatically substitute images, data and tables into a google doc. 

Version: 0.2.1

Generated: 2023-09-29T16:53:46.640578179
}

\details{
	
R library to perform limited interactions with google docs (and maybe one day slides)
in R via the Java API library. The purpose being to support google docs as a
platform for interactive development and documentation of data analysis in R for scientific
publication, although it is not limited to this purpose. The workflow supported is a parallel documentation and analysis
where a team of people are working collaboratively on documentation, whilst at the same time analysis
is being performed and results updated repeatedly as a result of new data. In this environment updating
numeric results, tabular data and figures in word documents manually becomes annoying. With roogledocs
you can automate this a bit like a RMarkdown document, but with the added benefit that the content
can be updated independently of the analysis, by the wider team.
}

\examples{
## -----------------------------------
## Construct new instance of RoogleDocs
## -----------------------------------
\dontrun{
J = roogledocs::JavaApi$get()
# appropriate parameter values must be provided
instance = J$RoogleDocs$new(tokenDirectory, disabled)
}

## -----------------------------------
## Method `RoogleDocs$enable()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$enable()
}

## -----------------------------------
## Method `RoogleDocs$disable()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$disable()
}

## -----------------------------------
## Method `RoogleDocs$withDocument(shareUrlOrDocId)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$withDocument(shareUrlOrDocId)
}

## -----------------------------------
## Method `RoogleDocs$findOrCreateDocument(title)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$findOrCreateDocument(title)
}

## -----------------------------------
## Method `RoogleDocs$findOrCloneTemplate(title, templateUri)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$findOrCloneTemplate(title, templateUri)
}

## -----------------------------------
## Method `RoogleDocs$tagsDefined()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$tagsDefined()
}

## -----------------------------------
## Method `RoogleDocs$updateTaggedText(text, tagName)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$updateTaggedText(text, tagName)
}

## -----------------------------------
## Method `RoogleDocs$updateTaggedImage(absoluteFilePath, tagName, dpi)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$updateTaggedImage(absoluteFilePath, tagName, dpi)
}

## -----------------------------------
## Method `RoogleDocs$revertTags()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$revertTags()
}

## -----------------------------------
## Method `RoogleDocs$removeTags(confirm)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$removeTags(confirm)
}

## -----------------------------------
## Method `RoogleDocs$updateTable(longFormatTable, tableIndex, colWidths, tableWidthInches)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$updateTable(longFormatTable, tableIndex, colWidths, tableWidthInches)
}

## -----------------------------------
## Method `RoogleDocs$updateFigure(absoluteFilePath, figureIndex, dpi)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$updateFigure(absoluteFilePath, figureIndex, dpi)
}

## -----------------------------------
## Method `RoogleDocs$saveAsPdf(absoluteFilePath, uploadCopy)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$saveAsPdf(absoluteFilePath, uploadCopy)
}

## -----------------------------------
## Method `RoogleDocs$uploadSupplementaryFiles(absoluteFilePath, overwrite, duplicate)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$uploadSupplementaryFiles(absoluteFilePath, overwrite, duplicate)
}

## -----------------------------------
## Method `RoogleDocs$appendText(text, style)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$appendText(text, style)
}

## -----------------------------------
## Method `RoogleDocs$appendFormattedParagraph(formattedTextDf)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$appendFormattedParagraph(formattedTextDf)
}

## -----------------------------------
## Method `RoogleDocs$updateCitations(bibTex, citationStyle)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$updateCitations(bibTex, citationStyle)
}
}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$RoogleDocs$new(tokenDirectory, disabled)}}
		}
	}
	\subsection{Class methods}{
		\itemize{
			\item \href{#method-enable}{\code{instance$enable()}}
			\item \href{#method-disable}{\code{instance$disable()}}
			\item \href{#method-withDocument}{\code{instance$withDocument(shareUrlOrDocId)}}
			\item \href{#method-findOrCreateDocument}{\code{instance$findOrCreateDocument(title)}}
			\item \href{#method-findOrCloneTemplate}{\code{instance$findOrCloneTemplate(title, templateUri)}}
			\item \href{#method-tagsDefined}{\code{instance$tagsDefined()}}
			\item \href{#method-updateTaggedText}{\code{instance$updateTaggedText(text, tagName)}}
			\item \href{#method-updateTaggedImage}{\code{instance$updateTaggedImage(absoluteFilePath, tagName, dpi)}}
			\item \href{#method-revertTags}{\code{instance$revertTags()}}
			\item \href{#method-removeTags}{\code{instance$removeTags(confirm)}}
			\item \href{#method-updateTable}{\code{instance$updateTable(longFormatTable, tableIndex, colWidths, tableWidthInches)}}
			\item \href{#method-updateFigure}{\code{instance$updateFigure(absoluteFilePath, figureIndex, dpi)}}
			\item \href{#method-saveAsPdf}{\code{instance$saveAsPdf(absoluteFilePath, uploadCopy)}}
			\item \href{#method-uploadSupplementaryFiles}{\code{instance$uploadSupplementaryFiles(absoluteFilePath, overwrite, duplicate)}}
			\item \href{#method-appendText}{\code{instance$appendText(text, style)}}
			\item \href{#method-appendFormattedParagraph}{\code{instance$appendFormattedParagraph(formattedTextDf)}}
			\item \href{#method-updateCitations}{\code{instance$updateCitations(bibTex, citationStyle)}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Constructor \code{new()}}{
Create a Roogledocs object for managing the interaction.

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to `.tokenDirectory()`)}{ - (java expects a String)}
				\item{disabled a flag to switch roogledocs off (on a document by document basis, for testing or development. This can be set globally with `options('roogledocs.disabled'=TRUE)` - (defaulting to `getOption('roogledocs.disabled',FALSE)`)}{ - (java expects a boolean)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
the new R6 RoogleDocs object
		}

	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-enable"></a>}}
	
	\subsection{Method \code{enable()}}{
Enables roogledocs method calls for this document. 

It is likely one of `withDocument()`, `findOrCreateDocument()` or `findOrCloneTemplate()` methods will be needed to specify the document.

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$enable()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-disable"></a>}}
	
	\subsection{Method \code{disable()}}{
Disables roogledocs temporarily for this document. 

While disabled all calls to roogledocs will silently abort.

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$disable()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-withDocument"></a>}}
	
	\subsection{Method \code{withDocument()}}{
Select a document by its share url or id.

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$withDocument(shareUrlOrDocId)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{shareUrlOrDocId the url from clicking a share button in google docs or an id from searchForDocuments() method}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-findOrCreateDocument"></a>}}
	
	\subsection{Method \code{findOrCreateDocument()}}{
Search for a document by name or create one if missing.

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$findOrCreateDocument(title)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{title a document title. If there is an exact match in google drive then that document will be used 
otherwise a new one will be created.}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-findOrCloneTemplate"></a>}}
	
	\subsection{Method \code{findOrCloneTemplate()}}{
Get a document by name or create one from a template if missing.

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$findOrCloneTemplate(title, templateUri)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{title a document title. If there is an exact match in google drive then that document will be used
otherwise a new one will be created.}{ - (java expects a String)}
				\item{templateUri the share link (or document id) of a template google document}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-tagsDefined"></a>}}
	
	\subsection{Method \code{tagsDefined()}}{
List all tags

Finds tags defined in the current document

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$tagsDefined()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RDataframe: a dataframe containing tag and count columns
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-updateTaggedText"></a>}}
	
	\subsection{Method \code{updateTaggedText()}}{
Relace tags for text

Substitutes all occurrences of {{tag-name}} with the text parameter.

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$updateTaggedText(text, tagName)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{text the value to replace the tag with (e.g. a result from analysis) (cannot be empty)}{ - (java expects a String)}
				\item{tagName the tag name - (defaulting to `deparse(substitute(text))`)}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-updateTaggedImage"></a>}}
	
	\subsection{Method \code{updateTaggedImage()}}{
Replace a tag with an image.

Substitutes all occurrences of {{tag-name}} with an image from the local storage. There are limited circumstances
in which using this is a good idea. It will almost always be better to use `updateFigure()` to insert an image
by index. If you choose to ignore this warning, beware combining this with `updateFigure()` as potentially the figure indexes may
change dynamically.

The image is uploaded to your google drive as a temporary file, and briefly made publically readable. From there it is inserted into the 
google doc, and one completed the temporary file deleted from your google drive.

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$updateTaggedImage(absoluteFilePath, tagName, dpi)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{absoluteFilePath a file path to an png image file.}{ - (java expects a String)}
				\item{tagName the tag name - (defaulting to `deparse(substitute(absoluteFilePath))`)}{ - (java expects a String)}
				\item{dpi the dots per inch of the image in the document (defaults to 300) - (defaulting to `300`)}{ - (java expects a double)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-revertTags"></a>}}
	
	\subsection{Method \code{revertTags()}}{
Revert tagged text and images.

Remove all tagged text and images inserted by roogledocs and returns the bare document the tags in place. This does not affect figures and tables inserted by index (i.e. without tags) 
This is needed if content is being moved around as cut and paste of tagged content unfortunately removes the internal named range of the tag.

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$revertTags()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-removeTags"></a>}}
	
	\subsection{Method \code{removeTags()}}{
Remove all tags

Finds tags defined in the current document and removes them. This 
cannot be undone.

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$removeTags(confirm)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{confirm - This action must be confirmed by passing `true` as cannot be undone. - (defaulting to `(menu(c('Yes','No'), title = 'Are you...`)}{ - (java expects a boolean)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-updateTable"></a>}}
	
	\subsection{Method \code{updateTable()}}{
Update or insert a formatted table into the document. 

The table and formatting are described in a dataframe the format of which is documented in the as.long_format_table() method.

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$updateTable(longFormatTable, tableIndex, colWidths, tableWidthInches)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{longFormatTable A dataframe consisting of the table content and formatting indexed by row and column. at a minimum this should have columns label,row,col, but may also include
rowSpan,colSpan,fillColour, leftBorderWeight, rightBorderWeight, topBorderWeight, bottomBorderWeight, alignment (START,CENTER,END), valignment (TOP,MIDDLE,BOTTOM), fontName, fontFace, fontSize.}{ - (java expects a RDataframe)}
				\item{tableIndex what is the table index in the document? This can be left out for a new table at the end of the document. - (defaulting to `-1`)}{ - (java expects a int)}
				\item{colWidths A vector including the relative length of each column. This can be left out if longFormatTable comes from as.long_format_table - (defaulting to `attr(longFormatTable,'colWidths')`)}{ - (java expects a RNumericVector)}
				\item{tableWidthInches The final width of the table in inches (defaults to a size that fits in A4 page with margins) - (defaulting to `6.2`)}{ - (java expects a RNumeric)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-updateFigure"></a>}}
	
	\subsection{Method \code{updateFigure()}}{
Update or insert a figure in the document from a locally stored PNG.

This function uploads the image into a temporary file onto your Google Drive, and makes it briefly publically readable. From there inserts it into the 
google document. Once this is complete the temporary google drive copy of the image is deleted.

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$updateFigure(absoluteFilePath, figureIndex, dpi)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{absoluteFilePath a file path to an png image file (only png is supported at this point).}{ - (java expects a String)}
				\item{figureIndex what is the figure index in the document? (This only counts inline images - and ignores absolutely positioned ones). leave out for a new image at the end of the document. - (defaulting to `-1`)}{ - (java expects a int)}
				\item{dpi the dots per inch of the image in the document (defaults to 300). the final size of the image in the doc will be determined by the image file dimensions and the dpi. - (defaulting to `300`)}{ - (java expects a double)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-saveAsPdf"></a>}}
	
	\subsection{Method \code{saveAsPdf()}}{
Save the document as a PDF

Saves a snapshot of the current google doc with `roogledocs` links removed as a pdf to a local drive. 
This is mainly intended for snapshotting the current state of the document. For final export once all
analysis is complete it may be preferable to call `doc$removeTags()` and manually export the output

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$saveAsPdf(absoluteFilePath, uploadCopy)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{absoluteFilePath - a file path to save the pdf.}{ - (java expects a String)}
				\item{uploadCopy place a copy of the downloaded pdf back onto google drive in the same folder as the document
  for example for keeping submitted versions of a updated document. This will overwrite files of the same name in the 
  google drive directory. - (defaulting to `FALSE`)}{ - (java expects a boolean)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-uploadSupplementaryFiles"></a>}}
	
	\subsection{Method \code{uploadSupplementaryFiles()}}{
Upload a file into the same directory as the document.

This allow you to load e.g. a supplementary file, or the pdf of an image file or a docx/html version of a table
into google drive into the same directory as the document you are editing. This is handy for organising all the files
for a journal submission in one place. Any kind of file can be loaded, and the mimetype will be detected. Normal Google Drive rules 
for uploads will be triggered at this point. As google drive can have multiple files with the same name
the behaviour if the file already exists is slightly complex, with `overwrite` and `duplicate` options.

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$uploadSupplementaryFiles(absoluteFilePath, overwrite, duplicate)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{absoluteFilePath - a file path to upload.}{ - (java expects a String)}
				\item{overwrite - if matching file(s) are found in the target, delete them before uploading the new one. - (defaulting to `FALSE`)}{ - (java expects a boolean)}
				\item{duplicate - if matching file(s) are found in the target, upload this new file anyway, creating duplicate names in the folder. - (defaulting to `FALSE`)}{ - (java expects a boolean)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-appendText"></a>}}
	
	\subsection{Method \code{appendText()}}{
Append text to the document with optional paragraph styling. If you run text blocks into each other without newlines the whole resulting paragraph will be styled. You 
would normally not want this so it is up to you to end paragraphs with a new line character, before changing styles.

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$appendText(text, style)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{text - a single string with the text to append which may include newlines}{ - (java expects a RCharacter)}
				\item{style - one of NORMAL_TEXT, TITLE, SUBTITLE, HEADING_1, ... HEADING_6 - (defaulting to `'NORMAL_TEXT'`)}{ - (java expects a RCharacter)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-appendFormattedParagraph"></a>}}
	
	\subsection{Method \code{appendFormattedParagraph()}}{
Append a new paragraph, with text from the 'label' column with optional formating in the other columns.

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$appendFormattedParagraph(formattedTextDf)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{formattedTextDf - a data frame containing the columns label, and optionally: link (as a URL), fontName, fontFace, fontSize.}{ - (java expects a RDataframe)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-updateCitations"></a>}}
	
	\subsection{Method \code{updateCitations()}}{
Update citation tags in the document. 

A citation tag is like this {{cite:challen2020;danon2021}}. The ids

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$updateCitations(bibTex, citationStyle)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{bibTex - a string containing the bibtex}{ - (java expects a RCharacter)}
				\item{citationStyle - the CSL specification - (defaulting to `'ieee'`)}{ - (java expects a RCharacter)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
R6 RoogleDocs object: itself - a fluent method
		}

	}
}